signoz/opentelemetry-demo/cart/requirements.txt
flask
requests
opentelemetry-distro
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-system-metrics
